+++ metasyn/demo/__init__.py
@@ -1,5 +1,5 @@
 """Package including demo datasets for tutorials."""
 
-from metasynth.demo.dataset import demo_file
+from metasyn.demo.dataset import demo_file
 
 __all__ = ["demo_file"]
