+++ metasyn/distribution/legacy/__init__.py
@@ -1,0 +1,7 @@
+"""Legacy distributions that will be removed in the future."""
+
+from metasyn.distribution.legacy.regex import RegexDistribution
+from metasyn.distribution.legacy.regex import UniqueRegexDistribution
+
+
+__all__ = ["RegexDistribution", "UniqueRegexDistribution"]
