+++ metasyn/distribution/legacy/regex/__init__.py
@@ -1,6 +1,6 @@
 """Package for fitting regexes to strings."""
 
-from metasynth.distribution.legacy.regex.base import (RegexDistribution,
+from metasyn.distribution.legacy.regex.base import (RegexDistribution,
                                                       UniqueRegexDistribution)
 
 __all__ = ["RegexDistribution", "UniqueRegexDistribution"]
