--- metasynth/distribution/legacy/__init__.py
@@ -1,7 +1,0 @@
-"""Legacy distributions that will be removed in the future."""
-
-from metasynth.distribution.legacy.regex import RegexDistribution
-from metasynth.distribution.legacy.regex import UniqueRegexDistribution
-
-
-__all__ = ["RegexDistribution", "UniqueRegexDistribution"]
