+++ tests/test_dataset.py
@@ -6,9 +6,9 @@ import pandas as pd
 import polars as pl
 from pytest import mark
 
-from metasynth.dataset import MetaFrame
-from metasynth.var import MetaVar
-from metasynth.provider import get_distribution_provider
+from metasyn.dataset import MetaFrame
+from metasyn.var import MetaVar
+from metasyn.provider import get_distribution_provider
 
 
 dtypes = {
